+++ lib/Extension/LanguageServerHover/Twig/TwigFunctions.php
@@ -2,8 +2,9 @@
 
 namespace Phpactor\Extension\LanguageServerHover\Twig;
 
+use Phpactor\Extension\LanguageServerHover\Twig\Functions\TypeShortName;
+use Phpactor\Extension\LanguageServerHover\Twig\Functions\TypeType;
 use Phpactor\WorseReflection\Core\Type;
-use Phpactor\WorseReflection\Core\Type\ReflectedClassType;
 use Phpactor\WorseReflection\TypeUtil;
 use Twig\Environment;
 use Twig\TwigFunction;
@@ -12,17 +13,12 @@ final class TwigFunctions
 {
     public static function add(Environment $env): Environment
     {
-        $env->addFunction(new TwigFunction('typeShortName', function (Type $type) {
-            if ($type instanceof ReflectedClassType) {
-                return TypeUtil::toLocalType($type);
-            }
-
-            return $type->__toString();
-        }));
+        $env->addFunction(new TwigFunction('typeShortName', new TypeShortName()));
 
         $env->addFunction(new TwigFunction('typeDefined', function (Type $type) {
             return TypeUtil::isDefined($type);
         }));
+        $env->addFunction(new TwigFunction('typeType', new TypeType()));
 
         return $env;
     }